ai-brain-mri-analyzer/backend/requirements.txt
Flask
Flask-Cors
torch
torchvision
numpy
Pillow
scikit-image
opencv-python
matplotlib